requirements.txt
beautifulsoup4==4.12.1
black==23.9.1
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
db-dtypes==1.1.1
fake-useragent==1.1.3
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
importlib-resources==5.12.0
lxml==4.9.2
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.1
pandas==2.0.0
pathspec==0.11.2
platformdirs==3.10.0
proto-plus==1.22.3
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
soupsieve==2.4
tzdata==2023.3
urllib3==1.26.15
zipp==3.15.0
